// models/personalData.ts
export interface PersonalData {
    id: string;
    name: string;
    firstName: string;
    lastName: string;
    phone?: string;
    email?: string;
    web?: string;
    home?: string;
    content: string;
    twitter?: string;
    github?: string;
    stackoverflow?: string;
    picture?: string;
}